Added new methods (decorate, test decorate etc)

+++ facade/src/main/java/mvcmodel/Model.java
@@ -7,12 +7,12 @@ import java.util.List;
  * Created by RENT on 2017-02-25.
  */
 public class Model {
-    public static List<Integer> values = new ArrayList<>();
+    public static List<Long> values = new ArrayList<>();
 
-    public static List<Integer> getValues() {
+    public static List<Long> getValues() {
         return values;
     }
-    public static void addValue(int newValue) {
+    public static void addValue(long newValue) {
         values.add(newValue);
     }
 }
